[safetensors] Extract parameter analysis
