+++ packages/hub/src/lib/parse-safetensors-metadata.spec.ts
@@ -561,89 +561,6 @@ describe("parseSafetensorsMetadata", () => {
 			);
 		});
 
-		it("requires a coherent pair of positive integer config values", async () => {
-			const header = {
-				__metadata__: { format: "pt" },
-				"model.embed_tokens.weight": { dtype: "F32", shape: [20], data_offsets: [0, 0] },
-				"model.layers.0.mlp.experts.weight": { dtype: "F32", shape: [4, 10], data_offsets: [0, 0] },
-			};
-			const parseWithConfig = (config: Record<string, unknown>) =>
-				parseSafetensorsMetadata({
-					repo: "some-user/synthetic-config-moe",
-					path: "model.safetensors",
-					computeParametersCount: true,
-					fetch: fetchForFile(header, 0, config),
-				});
-
-			const [mixedSources, fractional, nested, switchStyle, dbrxStyle] = await Promise.all([
-				parseWithConfig({ num_experts_per_tok: 1, text_config: { num_local_experts: 4 } }),
-				parseWithConfig({ num_experts_per_tok: 2, num_local_experts: 4.5 }),
-				parseWithConfig({
-					num_experts_per_tok: "invalid",
-					num_local_experts: 4,
-					text_config: { num_experts_per_tok: 2, num_local_experts: 4 },
-				}),
-				parseWithConfig({ num_selected_experts: 1, num_experts: 4 }),
-				parseWithConfig({ ffn_config: { moe_top_k: 2, moe_num_experts: 4 } }),
-			]);
-
-			assert.strictEqual(mixedSources.moe, undefined);
-			assert.strictEqual(fractional.moe, undefined);
-			assert.deepStrictEqual(nested.moe, {
-				numExperts: 4,
-				topK: 2,
-				perExpert: 10,
-				alwaysActive: 20,
-				active: 40,
-				hasSharedExpert: false,
-			});
-			assert.strictEqual(switchStyle.moe?.active, 30);
-			assert.strictEqual(dbrxStyle.moe?.active, 40);
-		});
-
-		it("requires a complete, in-range per-expert tensor set", async () => {
-			const parseWithExpertIds = (expertIds: number[]) =>
-				parseSafetensorsMetadata({
-					repo: "some-user/synthetic-per-expert-moe",
-					path: "model.safetensors",
-					computeParametersCount: true,
-					fetch: fetchForFile(
-						{
-							__metadata__: { format: "pt" },
-							"model.embed_tokens.weight": { dtype: "F32", shape: [11], data_offsets: [0, 0] },
-							"model.layers.0.mlp.router.weight": { dtype: "F32", shape: [4], data_offsets: [0, 0] },
-							"model.layers.0.mlp.shared_expert.weight": { dtype: "F32", shape: [7], data_offsets: [0, 0] },
-							"model.layers.0.mlp.shared_experts.weight": { dtype: "F32", shape: [5], data_offsets: [0, 0] },
-							...Object.fromEntries(
-								expertIds.map((expertId) => [
-									`model.layers.0.mlp.experts.${expertId}.weight`,
-									{ dtype: "F32", shape: [6], data_offsets: [0, 0] },
-								]),
-							),
-						},
-						0,
-						{ num_experts_per_tok: 2, num_local_experts: 4 },
-					),
-				});
-
-			const [complete, incomplete, outOfRange] = await Promise.all([
-				parseWithExpertIds([0, 1, 2, 3]),
-				parseWithExpertIds([0, 1, 2]),
-				parseWithExpertIds([0, 1, 2, 4]),
-			]);
-
-			assert.deepStrictEqual(complete.moe, {
-				numExperts: 4,
-				topK: 2,
-				perExpert: 6,
-				alwaysActive: 27,
-				active: 39,
-				hasSharedExpert: true,
-			});
-			assert.strictEqual(incomplete.moe, undefined);
-			assert.strictEqual(outOfRange.moe, undefined);
-		});
-
 		it("omits model-level MoE stats for an explicitly requested shard", async () => {
 			const fetch = fetchForFile(
 				{
